daemon: address PR #322 tenth-pass review feedback
Six new Copilot review comments on commit fa408bf, all applied:

- internal/socket/peercred*.go: change PeerCred.UID/GID from int32 to
  uint32. uid_t and gid_t are unsigned 32-bit on every platform we
  support (linux ucred.Uid/Gid, darwin xucred.Uid/Groups[]); narrowing
  into int32 wrapped UIDs above 2^31 to negative values and would have
  recorded corrupt peer attestation in receipts. Pipeline now formats
  the disclosure values via strconv.FormatUint.
  PID stays int32 — pid_t is signed.
- daemon.go + new umask_unix.go / umask_other.go: apply a restrictive
  process umask (0o027) before store.Open. SQLite creates WAL/SHM
  sidecars lazily — sometimes after tightenDBFiles has already run —
  so umask catches the new-file case at the source, while
  tightenDBFiles remains as belt-and-braces for files that already
  exist. Build-tagged so non-unix CI still compiles.
- daemon/README.md: document the CI coverage gap. None of the
  repository's GitHub Actions workflow path filters cover daemon/**,
  so daemon-only changes wouldn't trigger any vet/build/test job.
  AGENTS.md prohibits modifying CI workflows without explicit human
  review, so this PR documents the gap and points at the #236
  follow-up tracker; the maintainer can add a daemon.yml when
  appropriate.
- pipeline/build.go: validateFrame now requires non-empty ts_emit and
  parses it as RFC3339Nano. The daemon doesn't trust ts_emit for
  receipt timestamps (those come from p.Now()), but pinning the
  emitter contract here prevents a buggy emitter from silently
  shipping with malformed or absent timestamps. Updated table-test
  cases to (a) include a valid ts_emit where it isn't the field
  under test, and (b) cover the new "missing ts_emit" / "malformed
  ts_emit" rejection paths.

All sdk/go, daemon (unit + integration, workspace AND GOWORK=off),
and mcp-proxy tests pass under -race on linux/amd64.

Refs: #236, #322

+++ daemon/internal/socket/peercred_darwin.go
@@ -33,9 +33,11 @@ func capturePeer(conn *net.UnixConn) (PeerCred, error) {
 			inner = fmt.Errorf("LOCAL_PEERCRED: %w", err)
 			return
 		}
-		pc.UID = int32(xu.Uid)
+		// xu.Uid is uint32; xu.Groups elements are uint32. PeerCred.UID/GID
+		// are uint32 too — direct assignment, no narrowing.
+		pc.UID = xu.Uid
 		if xu.Ngroups > 0 {
-			pc.GID = int32(xu.Groups[0])
+			pc.GID = xu.Groups[0]
 		}
 
 		pid, err := unix.GetsockoptInt(int(fd), unix.SOL_LOCAL, unix.LOCAL_PEEREPID)
